sovereign_final/requirements.txt
streamlit>=1.35.0
psycopg2-binary>=2.9.9
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pandas>=2.0.0
apscheduler>=3.10.0
lxml>=5.0.0
pytz>=2024.1
openai>=1.30.0
